% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HEAT_SP}
\alias{HEAT_SP}
\title{HEAT SP}
\format{An dataframe with 14 observations and on the following two variables: \code{FData} and \code{Session}.}
\source{
\href{https://github.com/Debbiewm5/SPRE.R/blob/master/data/HEAT_LILLY_Heat.SP.14_2016\%202.csv}{A Github link to raw dataset}
}
\usage{
HEAT_SP
}
\description{
HEAT_SP data
}
\examples{
data("HEAT_SP")
}
\references{
Weissman-Miller, D., Miller, R. J., & Shotwell, M. P. (2017). Translational Research for Occupational Therapy: Using SPRE in Hippotherapy for Children with Developmental Disabilities. Occupational therapy international, 2017.
(\href{https://doi.org/10.1155/2017/2305402}{DOI})
}
\keyword{datasets}
