% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HEAT_stat}
\alias{HEAT_stat}
\title{HEAT stat}
\format{An dataframe with 14 observations and on the following two variables: \code{FData} and \code{Session}.}
\source{
\href{https://github.com/Debbiewm5/SPRE.R/blob/master/data/HEAT_LILLY_Heat.Stat.14_2016.csv}{A Github link to raw dataset}
}
\usage{
HEAT_stat
}
\description{
HEAT_stat data
}
\examples{
data("HEAT_stat")
}
\references{
Weissman-Miller, D., Miller, R. J., & Shotwell, M. P. (2017). Translational Research for Occupational Therapy: Using SPRE in Hippotherapy for Children with Developmental Disabilities. Occupational therapy international, 2017.
(\href{https://doi.org/10.1155/2017/2305402}{DOI})
}
\keyword{datasets}
