% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_probability}
\alias{plot_probability}
\title{plot probability}
\usage{
plot_probability(SPRE_object, SPRE_pred, xmin = 0,
  xmax = max(SPRE_pred$xpred, na.rm = TRUE))
}
\arguments{
\item{SPRE_object}{SPRE analysis output, an object of class 'SPRE'}

\item{SPRE_pred}{SPRE prediction output, an object of class 'SPRE_pred'}

\item{xmin}{lower bound of plot range}

\item{xmax}{upper bound of plot range}
}
\value{
output information to console and plots to device
}
\description{
plot probability
}
\details{
plot weibull distrubition of SPRE predictions
}
\examples{
skip_indices<-1:2
step_reg<-stepwise_regression(HEAT_stat$Session,HEAT_stat$FData,skip_indices)
SPRE_mod<-SPRE(step_reg,chptidx=11,lastidx=14)
kshape<-SPRE_mod$kshape
idxmaxFstat<-which.max(SPRE_mod$step_reg$fDist[,'Fstat'])
tau<-SPRE_mod$step_reg$fDist[idxmaxFstat,'x']
theta1<-HEAT_stat$FData[SPRE_mod$chptidx]
xpred<-7:46
SPRE_pred<-predict_SPRE(xpred=xpred,kshape=kshape,tau=tau,
                        theta1=theta1,predtype='rolling')
plot_probability(SPRE_mod,SPRE_pred)
}
\references{
Weissman-Miller, D., Shotwell, M.P. & Miller, R.J. (2012).  New single-subject and small-n design in occupational therapy:  Application to weight loss in obesity.  American Journal of Occupational Therapy, 66, 455-462. DOI: http://dx.doi.org/10.5014/ajot.2012.004788
Weissman-Miller, D. (2013). Novel point estimation from a Semiparametric Ratio Estimator (SPRE): Long-term health outcomes from short-term linear data, with application to weight loss in obesity. International Journal of Biostatistics, 9(2): 175-184 DOI: http://dx.doi.org/10.1515/ijb-2012-0049
Weissman-Miller, D., & Graham, K. C. (2015). Novel scale development for fear of falling and falls: Analysed using a Semiparametric Ratio Estimator (SPRE). International Journal of Statistics and Probability, 4(3), 161.
Weissman-Miller, D. (2016). On predicting survival in prostate cancer: using an extended maximum spacing method at the change point of the semiparametric ratio estimator (SPRE). International Journal of Statistics and Probability, 5(2), 19.
Weissman-Miller, D., Miller, R. J., & Shotwell, M. P. (2017). Translational Research for Occupational Therapy: Using SPRE in Hippotherapy for Children with Developmental Disabilities. Occupational therapy international, 2017.
}
\author{
Deborah Weissman-Miller
}
