% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stepwise_regression}
\alias{stepwise_regression}
\title{stepwise regression}
\usage{
stepwise_regression(x, y, skip_indices)
}
\arguments{
\item{x}{a vector of times/sessions}

\item{y}{a vector of responses}

\item{skip_indices}{indices where analyses are not performed}
}
\value{
an object of class 'SPRE_step_reg'
}
\description{
stepwise regression
}
\details{
generate a stepwise regression table from the data
}
\examples{
skip_indices<-1:2
step_reg<-stepwise_regression(HEAT_stat$Session,HEAT_stat$FData,skip_indices)
}
\references{
Weissman-Miller, D., Shotwell, M.P. & Miller, R.J. (2012).  New single-subject and small-n design in occupational therapy:  Application to weight loss in obesity.  American Journal of Occupational Therapy, 66, 455-462. DOI: http://dx.doi.org/10.5014/ajot.2012.004788
Weissman-Miller, D. (2013). Novel point estimation from a Semiparametric Ratio Estimator (SPRE): Long-term health outcomes from short-term linear data, with application to weight loss in obesity. International Journal of Biostatistics, 9(2): 175-184 DOI: http://dx.doi.org/10.1515/ijb-2012-0049
Weissman-Miller, D., & Graham, K. C. (2015). Novel scale development for fear of falling and falls: Analysed using a Semiparametric Ratio Estimator (SPRE). International Journal of Statistics and Probability, 4(3), 161.
Weissman-Miller, D. (2016). On predicting survival in prostate cancer: using an extended maximum spacing method at the change point of the semiparametric ratio estimator (SPRE). International Journal of Statistics and Probability, 5(2), 19.
Weissman-Miller, D., Miller, R. J., & Shotwell, M. P. (2017). Translational Research for Occupational Therapy: Using SPRE in Hippotherapy for Children with Developmental Disabilities. Occupational therapy international, 2017.
}
\author{
Deborah Weissman-Miller
}
